+++ examples/chapter5.rs
@@ -5,7 +5,7 @@ fn main() -> anyhow::Result<()> {
     let wall_z = 10.0;
     let wall_size = 7.0;
     let canvas_pixels = 100;
-    let pixel_size = wall_size / f64::from(canvas_pixels as u16);
+    let pixel_size = wall_size / f64::from(canvas_pixels);
     let half = wall_size / 2.0;
 
     let mut canvas = Canvas::new(canvas_pixels, canvas_pixels);
@@ -13,10 +13,10 @@ fn main() -> anyhow::Result<()> {
     let shape = sphere();
 
     for y in 0..canvas_pixels {
-        let world_y = half - pixel_size * f64::from(y as u16);
+        let world_y = half - pixel_size * f64::from(y);
 
         for x in 0..canvas_pixels {
-            let world_x = -half + pixel_size * f64::from(x as u16);
+            let world_x = -half + pixel_size * f64::from(x);
             let position = point(world_x, world_y, wall_z);
             let ray = ray(ray_origin, (position - ray_origin).normalize());
             let xs = shape.intersect(ray);
